+++ src/Products/PloneMeeting/utils.py
@@ -1722,7 +1722,8 @@ def checkMayQuickEdit(obj,
                       bypassWritePermissionCheck=False,
                       permission=ModifyPortalContent,
                       expression='',
-                      onlyForManagers=False):
+                      onlyForManagers=False,
+                      bypassMeetingClosedCheck=False):
     """ """
     from Products.PloneMeeting.content.meeting import Meeting
     tool = api.portal.get_tool('portal_plonemeeting')
@@ -1732,7 +1733,9 @@ def checkMayQuickEdit(obj,
     if (not onlyForManagers or (onlyForManagers and tool.isManager(obj))) and \
        (bypassWritePermissionCheck or member.has_permission(permission, obj)) and \
        (_evaluateExpression(obj, expression)) and \
-       (not (meeting and meeting.query_state() in Meeting.MEETINGCLOSEDSTATES) or
+       (not (not bypassMeetingClosedCheck and
+                meeting and
+                meeting.query_state() in Meeting.MEETINGCLOSEDSTATES) or
             tool.isManager(obj, realManagers=True)):
         res = True
     return res
@@ -1916,12 +1919,12 @@ def compute_item_roles_to_assign_to_suffixes(cfg, item_state, org=None):
     # states out of item validation (validated and following states)
     else:
         # every item validation suffixes get View access
-        # if item is in a decided state, we also give the Contributor
-        # role to every validation suffixes
-        item_is_decided = item_state in cfg.getItemDecidedStates()
+        # we also give the Contributor except to 'observers'
+        # so every editors roles get the "PloneMeeting: Add decision annex"
+        # permission that let add decision annex and write the MeetingItem.internalNotes
         for suffix in get_item_validation_wf_suffixes(cfg, org):
             given_roles = ['Reader']
-            if item_is_decided and suffix != 'observers':
+            if suffix != 'observers':
                 given_roles.append('Contributor')
             suffix_roles[suffix] = given_roles